Divided mapper into 2 files
- created folder mapper and divided the mapper by schema
